Standard deviation of the Mean Germination Time

This function calculates the standard deviation of the mean germination time

Usage

ger_SDG(evalName, data)

Arguments

evalName

Prefix of the names of the periods of evaluation.

data

The name of the data frame containing the data.

Value

It returns an vector with the values of Standard deviation of germination

Examples

library(GerminaR)
dfr <- prosopis
sdg <- ger_SDG(evalName = "D", data = dfr)
sdg
#>  [1] 0.5728554 0.4646702 0.4712121 0.3505098 0.1979487 0.2398979 0.0000000
#>  [8] 0.1414214 0.6144518 0.6144518 0.3282607 0.3703280 0.3030458 0.3703280
#> [15] 0.2740475 0.2398979 0.6944563 0.8741764 0.9290262 0.7142857 0.4545686
#> [22] 0.3730733 0.3766218 0.3505098 1.3113748 1.4433757 1.4438363 1.4869043
#> [29] 0.8417668 0.4932193 0.6144518 0.5089395 0.7831481 1.0300750 0.9630868
#> [36] 0.9906021 0.5983068 0.6631111 0.7462518 1.0174245 0.1979487 0.0000000
#> [43] 0.1414214 0.0000000 0.2398979 0.1979487 0.1414214 0.1414214 0.4797959
#> [50] 0.2446230 0.3136357 0.2446230 0.9301872 0.8551564 0.8660254 0.8461700
#> [57] 0.5609152 0.7067933 0.8769733 0.5000000 0.6824326 0.7986580 0.9128709
#> [64] 0.6942916 0.8545198 0.7576443 0.7216761 0.8615956 0.8164966 0.9831921
#> [71] 0.9574271 0.8944272 2.2010099 1.8378732 1.9436506 1.6624188 0.0000000
#> [78] 0.0000000 0.0000000 0.0000000