Summary of Germination indices

This function makes a data table with the result of germination indices for each experimental unit.

Usage,

ger_summary(SeedN, evalName, data)

Arguments

SeedN

Name of the column with the seed numbers

evalName

Prefix of the evaluation variable

data

The name of the data frame containing the data.

Value

Data frame with the summary values of germination variables.

Examples

library(GerminaR)
fb <- prosopis
smr <- ger_summary(SeedN = "seeds", evalName = "D", data = fb)
smr
#>    rep nacl temp seeds grs grp      mgt       mgr       gsp       unc
#> 1    1  0.0   25    50  50 100 1.280000 0.7812500  78.12500 0.9461447
#> 2    2  0.0   25    50  50 100 1.220000 0.8196721  81.96721 0.8157272
#> 3    3  0.0   25    50  50 100 1.320000 0.7575758  75.75758 0.9043815
#> 4    4  0.0   25    50  50 100 1.140000 0.8771930  87.71930 0.5842388
#> 5    1  0.0   30    50  50 100 1.040000 0.9615385  96.15385 0.2422922
#> 6    2  0.0   30    50  50 100 1.060000 0.9433962  94.33962 0.3274449
#> 7    3  0.0   30    50  50 100 1.000000 1.0000000 100.00000 0.0000000
#> 8    4  0.0   30    50  50 100 1.020000 0.9803922  98.03922 0.1414405
#> 9    1  0.5   25    50  50 100 1.900000 0.5263158  52.63158 1.0844751
#> 10   2  0.5   25    50  50 100 1.700000 0.5882353  58.82353 1.1985488
#> 11   3  0.5   25    50  50 100 1.880000 0.5319149  53.19149 0.5293609
#> 12   4  0.5   25    50  50 100 1.840000 0.5434783  54.34783 0.6343096
#> 13   1  0.5   30    50  50 100 1.100000 0.9090909  90.90909 0.4689956
#> 14   2  0.5   30    50  50 100 1.160000 0.8620690  86.20690 0.6343096
#> 15   3  0.5   30    50  50 100 1.080000 0.9259259  92.59259 0.4021792
#> 16   4  0.5   30    50  50 100 1.060000 0.9433962  94.33962 0.3274449
#> 17   1  1.0   25    50  48  96 2.666667 0.3750000  37.50000 1.4171327
#> 18   2  1.0   25    50  48  96 2.708333 0.3692308  36.92308 1.4081359
#> 19   3  1.0   25    50  47  94 2.531915 0.3949580  39.49580 1.5822405
#> 20   4  1.0   25    50  49  98 2.897959 0.3450704  34.50704 1.4950825
#> 21   1  1.0   30    50  49  98 1.959184 0.5104167  51.04167 0.9281698
#> 22   2  1.0   30    50  50 100 1.940000 0.5154639  51.54639 0.7050757
#> 23   3  1.0   30    50  48  96 1.833333 0.5454545  54.54545 0.6500224
#> 24   4  1.0   30    50  50 100 1.860000 0.5376344  53.76344 0.5842388
#> 25   1  1.5   25    50  47  94 5.382979 0.1857708  18.57708 1.6951591
#> 26   2  1.5   25    50  48  96 5.458333 0.1832061  18.32061 1.8213883
#> 27   3  1.5   25    50  48  96 5.479167 0.1825095  18.25095 1.6026878
#> 28   4  1.5   25    50  49  98 5.448980 0.1835206  18.35206 1.5043742
#> 29   1  1.5   30    50  50 100 3.160000 0.3164557  31.64557 1.5468954
#> 30   2  1.5   30    50  50 100 2.960000 0.3378378  33.78378 1.0302088
#> 31   3  1.5   30    50  50 100 2.900000 0.3448276  34.48276 1.2633065
#> 32   4  1.5   30    50  47  94 2.957447 0.3381295  33.81295 1.0697797
#> 33   1  2.0   25    50  47  94 6.680851 0.1496815  14.96815 1.6179042
#> 34   2  2.0   25    50  47  94 6.063830 0.1649123  16.49123 1.9947498
#> 35   3  2.0   25    50  46  92 6.695652 0.1493506  14.93506 1.9243519
#> 36   4  2.0   25    50  49  98 6.653061 0.1503067  15.03067 1.5071571
#> 37   1  2.0   30    50  46  92 4.326087 0.2311558  23.11558 1.2646502
#> 38   2  2.0   30    50  48  96 4.333333 0.2307692  23.07692 1.3775500
#> 39   3  2.0   30    50  47  94 4.446809 0.2248804  22.48804 1.6034362
#> 40   4  2.0   30    50  47  94 4.446809 0.2248804  22.48804 1.9685404
#> 41   1  0.0   35    50  50 100 1.040000 0.9615385  96.15385 0.2422922
#> 42   2  0.0   35    50  50 100 1.000000 1.0000000 100.00000 0.0000000
#> 43   3  0.0   35    50  50 100 1.020000 0.9803922  98.03922 0.1414405
#> 44   4  0.0   35    50  50 100 1.000000 1.0000000 100.00000 0.0000000
#> 45   1  0.0   40    50  50 100 1.060000 0.9433962  94.33962 0.3274449
#> 46   2  0.0   40    50  50 100 1.040000 0.9615385  96.15385 0.2422922
#> 47   3  0.0   40    50  50 100 1.020000 0.9803922  98.03922 0.1414405
#> 48   4  0.0   40    50  50 100 1.020000 0.9803922  98.03922 0.1414405
#> 49   1  0.5   35    50  50 100 1.120000 0.8928571  89.28571 0.3274449
#> 50   2  0.5   35    50  48  96 1.062500 0.9411765  94.11765 0.3372901
#> 51   3  0.5   35    50  50 100 1.060000 0.9433962  94.33962 0.2822922
#> 52   4  0.5   35    50  48  96 1.062500 0.9411765  94.11765 0.3372901
#> 53   1  0.5   40    50  48  96 2.333333 0.4285714  42.85714 1.1228074
#> 54   2  0.5   40    50  49  98 2.346939 0.4260870  42.60870 1.3718323
#> 55   3  0.5   40    50  48  96 2.375000 0.4210526  42.10526 1.2987949
#> 56   4  0.5   40    50  47  94 2.255319 0.4433962  44.33962 1.4883676
#> 57   1  1.0   35    50  49  98 1.653061 0.6049383  60.49383 1.1796780
#> 58   2  1.0   35    50  48  96 1.729167 0.5783133  57.83133 1.4531143
#> 59   3  1.0   35    50  46  92 2.173913 0.4600000  46.00000 1.5098718
#> 60   4  1.0   35    50  49  98 1.714286 0.5833333  58.33333 1.0214779
#> 61   1  1.0   40    50  50 100 2.940000 0.3401361  34.01361 1.4015264
#> 62   2  1.0   40    50  48  96 2.520833 0.3966942  39.66942 1.6336284
#> 63   3  1.0   40    50  49  98 2.714286 0.3684211  36.84211 1.8160410
#> 64   4  1.0   40    50  50 100 2.740000 0.3649635  36.49635 1.1510457
#> 65   1  1.5   35    50  50 100 3.380000 0.2958580  29.58580 1.7518407
#> 66   2  1.5   35    50  48  96 3.354167 0.2981366  29.81366 1.5487081
#> 67   3  1.5   35    50  50 100 3.360000 0.2976190  29.76190 1.5566689
#> 68   4  1.5   35    50  49  98 3.387755 0.2951807  29.51807 1.6343886
#> 69   1  1.5   40    50   6  12 3.333333 0.3000000  30.00000 1.4591479
#> 70   2  1.5   40    50   6  12 3.166667 0.3157895  31.57895 1.4591479
#> 71   3  1.5   40    50   4   8 3.250000 0.3076923  30.76923 1.5000000
#> 72   4  1.5   40    50   5  10 3.400000 0.2941176  29.41176 1.3709506
#> 73   1  2.0   35    50  10  20 6.800000 0.1470588  14.70588 2.6464393
#> 74   2  2.0   35    50  10  20 7.400000 0.1351351  13.51351 1.4854753
#> 75   3  2.0   35    50   9  18 6.555556 0.1525424  15.25424 1.3516441
#> 76   4  2.0   35    50  11  22 7.181818 0.1392405  13.92405 2.1626441
#> 77   1  2.0   40    50   0   0      NaN       NaN       NaN 0.0000000
#> 78   2  2.0   40    50   0   0      NaN       NaN       NaN 0.0000000
#> 79   3  2.0   40    50   0   0      NaN       NaN       NaN 0.0000000
#> 80   4  2.0   40    50   0   0      NaN       NaN       NaN 0.0000000
#>           syn        vgt       sdg      cvg
#> 1  0.63020408 0.32816327 0.5728554 44.75433
#> 2  0.66612245 0.21591837 0.4646702 38.08772
#> 3  0.55591837 0.22204082 0.4712121 35.69788
#> 4  0.75428571 0.12285714 0.3505098 30.74648
#> 5  0.92163265 0.03918367 0.1979487 19.03353
#> 6  0.88489796 0.05755102 0.2398979 22.63188
#> 7  1.00000000 0.00000000 0.0000000  0.00000
#> 8  0.96000000 0.02000000 0.1414214 13.86484
#> 9  0.58122449 0.37755102 0.6144518 32.33957
#> 10 0.48000000 0.37755102 0.6144518 36.14422
#> 11 0.78448980 0.10775510 0.3282607 17.46068
#> 12 0.72571429 0.13714286 0.3703280 20.12652
#> 13 0.81632653 0.09183673 0.3030458 27.54961
#> 14 0.72571429 0.13714286 0.3703280 31.92483
#> 15 0.84979592 0.07510204 0.2740475 25.37477
#> 16 0.88489796 0.05755102 0.2398979 22.63188
#> 17 0.38652482 0.48226950 0.6944563 26.04211
#> 18 0.40514184 0.76418440 0.8741764 32.27728
#> 19 0.39222942 0.86308973 0.9290262 36.69263
#> 20 0.36224490 0.51020408 0.7142857 24.64789
#> 21 0.64795918 0.20663265 0.4545686 23.20194
#> 22 0.74612245 0.13918367 0.3730733 19.23058
#> 23 0.71631206 0.14184397 0.3766218 20.54301
#> 24 0.75428571 0.12285714 0.3505098 18.84461
#> 25 0.30712303 1.71970398 1.3113748 24.36151
#> 26 0.29521277 2.08333333 1.4433757 26.44352
#> 27 0.34485816 2.08466312 1.4438363 26.35138
#> 28 0.35374150 2.21088435 1.4869043 27.28776
#> 29 0.48081633 0.70857143 0.8417668 26.63819
#> 30 0.59918367 0.24326531 0.4932193 16.66282
#> 31 0.50775510 0.37755102 0.6144518 21.18799
#> 32 0.57909343 0.25901943 0.5089395 17.20875
#> 33 0.39037928 0.61332100 0.7831481 11.72228
#> 34 0.26734505 1.06105458 1.0300750 16.98720
#> 35 0.29855072 0.92753623 0.9630868 14.38376
#> 36 0.40816327 0.98129252 0.9906021 14.88942
#> 37 0.44057971 0.35797101 0.5983068 13.83021
#> 38 0.39982270 0.43971631 0.6631111 15.30256
#> 39 0.36077706 0.55689177 0.7462518 16.78174
#> 40 0.28029602 1.03515264 1.0174245 22.87988
#> 41 0.92163265 0.03918367 0.1979487 19.03353
#> 42 1.00000000 0.00000000 0.0000000  0.00000
#> 43 0.96000000 0.02000000 0.1414214 13.86484
#> 44 1.00000000 0.00000000 0.0000000  0.00000
#> 45 0.88489796 0.05755102 0.2398979 22.63188
#> 46 0.92163265 0.03918367 0.1979487 19.03353
#> 47 0.96000000 0.02000000 0.1414214 13.86484
#> 48 0.96000000 0.02000000 0.1414214 13.86484
#> 49 0.88489796 0.23020408 0.4797959 42.83892
#> 50 0.88031915 0.05984043 0.2446230 23.02334
#> 51 0.92081633 0.09836735 0.3136357 29.58827
#> 52 0.88031915 0.05984043 0.2446230 23.02334
#> 53 0.50620567 0.86524823 0.9301872 39.86517
#> 54 0.42517007 0.73129252 0.8551564 36.43710
#> 55 0.45744681 0.75000000 0.8660254 36.46423
#> 56 0.36725254 0.71600370 0.8461700 37.51886
#> 57 0.46768707 0.31462585 0.5609152 33.93191
#> 58 0.37322695 0.49955674 0.7067933 40.87479
#> 59 0.35458937 0.76908213 0.8769733 40.34077
#> 60 0.53826531 0.25000000 0.5000000 29.16667
#> 61 0.47183673 0.46571429 0.6824326 23.21199
#> 62 0.36258865 0.63785461 0.7986580 31.68230
#> 63 0.30527211 0.83333333 0.9128709 33.63209
#> 64 0.59102041 0.48204082 0.6942916 25.33911
#> 65 0.32897959 0.73020408 0.8545198 25.28165
#> 66 0.40780142 0.57402482 0.7576443 22.58815
#> 67 0.37387755 0.52081633 0.7216761 21.47845
#> 68 0.36309524 0.74234694 0.8615956 25.43264
#> 69 0.26666667 0.66666667 0.8164966 24.49490
#> 70 0.26666667 0.96666667 0.9831921 31.04817
#> 71 0.16666667 0.91666667 0.9574271 29.45930
#> 72 0.30000000 0.80000000 0.8944272 26.30668
#> 73 0.08888889 4.84444444 2.2010099 32.36779
#> 74 0.31111111 3.37777778 1.8378732 24.83612
#> 75 0.36111111 3.77777778 1.9436506 29.64891
#> 76 0.16363636 2.76363636 1.6624188 23.14760
#> 77        NaN 0.00000000 0.0000000      NaN
#> 78        NaN 0.00000000 0.0000000      NaN
#> 79        NaN 0.00000000 0.0000000      NaN
#> 80        NaN 0.00000000 0.0000000      NaN