Variance of the Mean Germination Time

This function calculates the variance of the mean germination time.

Usage

ger_VGT(evalName, data)

Arguments

evalName

Prefix of the names of the periods of evaluation.

data

The name of the data frame containing the data.

Value

It returns an vector with the values of Variance of Germination

Examples

library(GerminaR)
dfr <- prosopis
vgt <- ger_VGT(evalName = "D", data = dfr)
vgt
#>  [1] 0.32816327 0.21591837 0.22204082 0.12285714 0.03918367 0.05755102
#>  [7] 0.00000000 0.02000000 0.37755102 0.37755102 0.10775510 0.13714286
#> [13] 0.09183673 0.13714286 0.07510204 0.05755102 0.48226950 0.76418440
#> [19] 0.86308973 0.51020408 0.20663265 0.13918367 0.14184397 0.12285714
#> [25] 1.71970398 2.08333333 2.08466312 2.21088435 0.70857143 0.24326531
#> [31] 0.37755102 0.25901943 0.61332100 1.06105458 0.92753623 0.98129252
#> [37] 0.35797101 0.43971631 0.55689177 1.03515264 0.03918367 0.00000000
#> [43] 0.02000000 0.00000000 0.05755102 0.03918367 0.02000000 0.02000000
#> [49] 0.23020408 0.05984043 0.09836735 0.05984043 0.86524823 0.73129252
#> [55] 0.75000000 0.71600370 0.31462585 0.49955674 0.76908213 0.25000000
#> [61] 0.46571429 0.63785461 0.83333333 0.48204082 0.73020408 0.57402482
#> [67] 0.52081633 0.74234694 0.66666667 0.96666667 0.91666667 0.80000000
#> [73] 4.84444444 3.37777778 3.77777778 2.76363636 0.00000000 0.00000000
#> [79] 0.00000000 0.00000000